+++ docker/build_and_export_qjs.sh
@@ -18,6 +18,7 @@ source "${script_dir}/../builder/env/qjs"
 # ARG_OPTIONAL_REPEATED([extra-dir],[e],[extra directory to add into package],[])
 # ARG_OPTIONAL_BOOLEAN([force-build-image],[],[force rebuilding docker image],[off])
 # ARG_OPTIONAL_BOOLEAN([verbose],[v],[enable verbose mode],[off])
+# ARG_OPTIONAL_BOOLEAN([upx],[u],[compress binaries using upx],[on])
 # ARG_POSITIONAL_SINGLE([qjs-version],[QuickJS version (ex: 2020-09-06)],[$default_qjs_version])
 # ARG_TYPE_GROUP_SET([arch],[type string],[arch],[x86_64,i686,armv7l,aarch64])
 # ARG_HELP([Build a static version of QuickJS (interpreter & compiler)])
@@ -51,7 +52,7 @@ arch()
 
 begins_with_short_option()
 {
-	local first_option all_short_options='paevh'
+	local first_option all_short_options='paevuh'
 	first_option="${1:0:1}"
 	test "$all_short_options" = "${all_short_options/$first_option/}" && return 1 || return 0
 }
@@ -67,12 +68,13 @@ _arg_ext_lib_version="$default_qjs_ext_lib_version"
 _arg_extra_dir=()
 _arg_force_build_image="off"
 _arg_verbose="off"
+_arg_upx="on"
 
 
 print_help()
 {
 	printf '%s\n' "Build a static version of QuickJS (interpreter & compiler)"
-	printf 'Usage: %s [-p|--packages-dir <arg>] [-a|--arch <type string>] [--(no-)ext-lib] [--ext-lib-version <arg>] [-e|--extra-dir <arg>] [--(no-)force-build-image] [-v|--(no-)verbose] [-h|--help] [<qjs-version>]\n' "$0"
+	printf 'Usage: %s [-p|--packages-dir <arg>] [-a|--arch <type string>] [--(no-)ext-lib] [--ext-lib-version <arg>] [-e|--extra-dir <arg>] [--(no-)force-build-image] [-v|--(no-)verbose] [-u|--(no-)upx] [-h|--help] [<qjs-version>]\n' "$0"
 	printf '\t%s\n' "<qjs-version>: QuickJS version (ex: 2020-09-06) (default: '$default_qjs_version')"
 	printf '\t%s\n' "-p, --packages-dir: directory where package will be exported (default: '$script_dir/../packages')"
 	printf '\t%s\n' "-a, --arch: target architecture. Can be one of: 'x86_64', 'i686', 'armv7l' and 'aarch64' (default: 'x86_64')"
@@ -81,6 +83,7 @@ print_help()
 	printf '\t%s\n' "-e, --extra-dir: extra directory to add into package (empty by default)"
 	printf '\t%s\n' "--force-build-image, --no-force-build-image: force rebuilding docker image (off by default)"
 	printf '\t%s\n' "-v, --verbose, --no-verbose: enable verbose mode (off by default)"
+	printf '\t%s\n' "-u, --upx, --no-upx: compress binaries using upx (on by default)"
 	printf '\t%s\n' "-h, --help: Prints help"
 }
 
@@ -153,6 +156,18 @@ parse_commandline()
 					{ begins_with_short_option "$_next" && shift && set -- "-v" "-${_next}" "$@"; } || die "The short option '$_key' can't be decomposed to ${_key:0:2} and -${_key:2}, because ${_key:0:2} doesn't accept value and '-${_key:2:1}' doesn't correspond to a short option."
 				fi
 				;;
+			-u|--no-upx|--upx)
+				_arg_upx="on"
+				test "${1:0:5}" = "--no-" && _arg_upx="off"
+				;;
+			-u*)
+				_arg_upx="on"
+				_next="${_key##-u}"
+				if test -n "$_next" -a "$_next" != "$_key"
+				then
+					{ begins_with_short_option "$_next" && shift && set -- "-u" "-${_next}" "$@"; } || die "The short option '$_key' can't be decomposed to ${_key:0:2} and -${_key:2}, because ${_key:0:2} doesn't accept value and '-${_key:2:1}' doesn't correspond to a short option."
+				fi
+				;;
 			-h|--help)
 				print_help
 				exit 0
@@ -207,10 +222,10 @@ assign_positional_args 1 "${_positionals[@]}"
 # vvv  PLACE YOUR CODE HERE  vvv
 
 # ensure version exist
-qjs_commit="${qjs_commits[$_arg_qjs_version]}"
+qjs_commit="${qjs_commits[${_arg_qjs_version}]}"
 if [ -z ${qjs_commit} ]
 then
-    _PRINT_HELP=yes die "QuickJS version '$_arg_qjs_version' is not supported"
+    _PRINT_HELP=yes die "QuickJS version '${_arg_qjs_version}' is not supported"
 fi
 
 _PRINT_HELP=no
@@ -218,9 +233,9 @@ _PRINT_HELP=no
 build_docker_image()
 {
     _flag_verbose=""
-    [ $_arg_verbose == "on" ] && _flag_verbose="-v"
+    [ ${_arg_verbose} == "on" ] && _flag_verbose="-v"
     _flag_force_build_image=""
-    [ $_arg_force_build_image == "on" ] && _flag_force_build_image="-f"
+    [ ${_arg_force_build_image} == "on" ] && _flag_force_build_image="-f"
 
     ${script_dir}/scripts/build_docker_image.sh ${_flag_verbose} -a ${_arg_arch} ${_flag_force_build_image} || return 1
 
@@ -232,19 +247,21 @@ build_and_export_package()
     _image_name="quickjs-cross-compiler:${_arg_arch}"
 
     _flag_verbose=""
-    [ $_arg_verbose == "on" ] && _flag_verbose="-v"
+    [ ${_arg_verbose} == "on" ] && _flag_verbose="-v"
     args_qjs_ext_lib=""
-    if [ $_arg_ext_lib == "on" ]
+    if [ ${_arg_ext_lib} == "on" ]
     then
-        args_qjs_ext_lib="--ext-lib --ext-lib-version $_arg_ext_lib_version"
+        args_qjs_ext_lib="--ext-lib --ext-lib-version ${_arg_ext_lib_version}"
     fi
     args_extra_dir=""
     for a in ${_arg_extra_dir[@]}
     do
         args_extra_dir="${args_extra_dir} -e ${a}"
     done
+    _flag_disable_upx=""
+    [ ${_arg_upx} == "off" ] && _flag_disable_upx="--no-upx"
 
-    _docker_cmd="./build_and_export_qjs.sh ${_flag_verbose} -a ${_arg_arch} ${args_qjs_ext_lib} ${args_extra_dir}"
+    _docker_cmd="./build_and_export_qjs.sh ${_flag_verbose} -a ${_arg_arch} ${args_qjs_ext_lib} ${args_extra_dir} ${_flag_disable_upx}"
 
     (mkdir -p ${_arg_packages_dir} && \
         docker run --rm \